Fix an issue in converting the secret to bytes
Instead of converting the secret to byte array from Base32 string, the
code used to convert the secret from ASCII string to byte array. Tests
used to pass because the secret in tests (taken from RFCs) hand-picked
to be the ASCII representation of the actual Base32 string. With this
change, the tests now use Base32 string for the secret.